pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "transaction-coordinator"
version = "0.1.0"
description = "AI-powered real estate transaction coordinator for California (LA / Beverly Hills)"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "anthropic>=0.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0",
    "docusign-esign>=3.20.0",
    "pymupdf>=1.23.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "jinja2>=3.1.0",
]

[project.scripts]
tc = "tc.cli:app"

[tool.setuptools.packages.find]
where = ["src"]
